fix: cleanup directories
