+++ packages/core-flows/src/customer/steps/index.ts
@@ -1,3 +1,3 @@
+export * from "./create-customers"
 export * from "./update-customers"
 export * from "./delete-customers"
-export * from "./create-customers"
