+++ packages/core-flows/src/customer/workflows/update-customers.ts
@@ -1,0 +1,22 @@
+import {
+  CustomerDTO,
+  CustomerUpdateableFields,
+  FilterableCustomerProps,
+} from "@medusajs/types"
+import { WorkflowData, createWorkflow } from "@medusajs/workflows-sdk"
+import { updateCustomersStep } from "../steps"
+
+type UpdateCustomersStepInput = {
+  selector: FilterableCustomerProps
+  update: CustomerUpdateableFields
+}
+
+type WorkflowInput = UpdateCustomersStepInput
+
+export const updateCustomersWorkflowId = "update-customers"
+export const updateCustomersWorkflow = createWorkflow(
+  updateCustomersWorkflowId,
+  (input: WorkflowData<WorkflowInput>): WorkflowData<CustomerDTO[]> => {
+    return updateCustomersStep(input)
+  }
+)
